Add AuditableEntity class with properties for tracking creation and modification timestamps
